pyproject.toml
[project]
name = "assessment"
version = "0.1.0"
description = "Send Money Agent — Google ADK + Protobuf + PostgreSQL 18 + Langfuse"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.28.0",
    "django>=6.0.3",
    "psycopg[binary]>=3.2.0",
    "asyncpg>=0.30.0",
    "googleapis-common-protos>=1.73.0",
    "protobuf>=6.33.0",
    "langfuse>=2.60.0",
    "opentelemetry-sdk>=1.33.0",
    "opentelemetry-exporter-otlp-proto-http>=1.33.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.0",
    "ruff>=0.15.8",
    "grpcio-tools>=1.73.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["backend"]
asyncio_mode = "auto"
DJANGO_SETTINGS_MODULE = "config.settings"

[tool.mypy]
mypy_path = "backend"
python_version = "3.14"
strict = true
ignore_missing_imports = true

[tool.ruff]
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
